moving denominations under commong
