+++ api/service/staking/service.go
@@ -19,9 +19,9 @@ import (
 	proto_common "github.com/harmony-one/harmony/api/proto"
 	"github.com/harmony-one/harmony/api/proto/message"
 	msg_pb "github.com/harmony-one/harmony/api/proto/message"
+	"github.com/harmony-one/harmony/common/denominations"
 	"github.com/harmony-one/harmony/contracts"
 	"github.com/harmony-one/harmony/core"
-	"github.com/harmony-one/harmony/core/denominations"
 	"github.com/harmony-one/harmony/core/types"
 	hmykey "github.com/harmony-one/harmony/internal/keystore"
 	"github.com/harmony-one/harmony/internal/utils"
