+++ common/denominations/denominations.go
@@ -1,0 +1,8 @@
+package denominations
+
+//ONE Units
+const (
+	Wei  = 1
+	GWei = 1e9
+	One  = 1e18
+)
